Backend/requirements.txt
langchain==0.2.10
langchain-google-genai==1.0.8
langchain-experimental==0.0.64
sqlalchemy==2.0.32

google-generativeai==0.7.2
python-dotenv==1.0.1
matplotlib==3.9.2
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
numpy==1.26.4
